move CreateNewTempSession() & GenerateTempAccessToken() into the app/auth package
